API CHANGE: Get the sysadm-bridge all finished up. Now to add the necessary bits to the server/client so we can try using it. NOTE: There is a special API/event for the bridge to announce which server/client connections are available (clients will see only servers, and vice versa) The syntax for this event is:
{
 "namespace":"events",
 "name":"bridge",
 "id":"",
 "args":{
   "available_connections":["ID1","ID2", etc..]
 }
}

Note that the available_connections are *all* the connections available at that time (it is not a difference from a previous state), so there may be both new ID's in the list and ID's which are no longer listed.

This event will get sent out any time a new connection/disconnection is made from the bridge which impacts the current connection (so a client will only get the event when a server connects/disconnects and vice versa).